renamed status package to event and moved out of ui
