+++ src/io/github/trdesilva/autorecorder/SettingsValidator.java
@@ -6,10 +6,10 @@
 package io.github.trdesilva.autorecorder;
 
 import com.google.inject.Inject;
-import io.github.trdesilva.autorecorder.ui.status.Event;
-import io.github.trdesilva.autorecorder.ui.status.EventProperty;
-import io.github.trdesilva.autorecorder.ui.status.EventQueue;
-import io.github.trdesilva.autorecorder.ui.status.EventType;
+import io.github.trdesilva.autorecorder.event.Event;
+import io.github.trdesilva.autorecorder.event.EventProperty;
+import io.github.trdesilva.autorecorder.event.EventQueue;
+import io.github.trdesilva.autorecorder.event.EventType;
 
 import java.io.File;
 import java.util.Collections;
