+++ src/io/github/trdesilva/autorecorder/ui/gui/record/ClippingPanel.java
@@ -12,9 +12,9 @@ import io.github.trdesilva.autorecorder.clip.ClipQueue;
 import io.github.trdesilva.autorecorder.ui.gui.Navigator;
 import io.github.trdesilva.autorecorder.ui.gui.VideoPlaybackPanel;
 import io.github.trdesilva.autorecorder.ui.gui.wrapper.DefaultPanel;
-import io.github.trdesilva.autorecorder.ui.status.Event;
-import io.github.trdesilva.autorecorder.ui.status.EventQueue;
-import io.github.trdesilva.autorecorder.ui.status.EventType;
+import io.github.trdesilva.autorecorder.event.Event;
+import io.github.trdesilva.autorecorder.event.EventQueue;
+import io.github.trdesilva.autorecorder.event.EventType;
 import net.miginfocom.swing.MigLayout;
 
 import javax.swing.JButton;
