+++ src/io/github/trdesilva/autorecorder/upload/youtube/YoutubeJobValidator.java
@@ -8,9 +8,9 @@ package io.github.trdesilva.autorecorder.upload.youtube;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import io.github.trdesilva.autorecorder.video.VideoFilenameValidator;
-import io.github.trdesilva.autorecorder.ui.status.Event;
-import io.github.trdesilva.autorecorder.ui.status.EventQueue;
-import io.github.trdesilva.autorecorder.ui.status.EventType;
+import io.github.trdesilva.autorecorder.event.Event;
+import io.github.trdesilva.autorecorder.event.EventQueue;
+import io.github.trdesilva.autorecorder.event.EventType;
 import io.github.trdesilva.autorecorder.upload.UploadJob;
 import io.github.trdesilva.autorecorder.upload.UploadJobValidator;
 import io.github.trdesilva.autorecorder.video.VideoListHandler;
